Convert ioRuntime to lazy val
By making ioRuntime a lazy val in the AsyncIOSpec the order in which the
trait are mixed in the trait does not matter as much as it is current.

Currently, by mixing AsyncIO trait BEFORE another trait/abstract class
with a executionContext the execution fails with a Null Pointer
Exception.